Add Gaussian filter
Use scikit-image's gaussing filter to smooth image. This filter must be
explicitly applied by the user if it is to be used as part of a
clustering operation. Some algorithms may be susceptible to smoothing as
edges in features tend to disappear. Missing edges that could be
important for labeling clustered areas can lead to false results or even
data loss.